Separate redis-py & coredis implementations
To allow for a backward compatible path for introducing
the asyncio implementation via redis-py, add an abstract
RedisBridge which can be selected at runtime based on
the ``implementation`` optional argument to RedisStorage
(which for now defaults to coredis)
