+++ pyproject.toml
@@ -8,3 +8,6 @@ quote-style = "double"
 indent-style = "space"
 skip-magic-trailing-comma = false
 line-ending = "auto"
+
+[tool.ruff.lint]
+typing-modules = ["limits.typing"]
