+++ requirements/storage/async-redis.txt
@@ -1,1 +1,1 @@
-redis>=4.2.0
+coredis>=3.4.0,<5
