+++ tests/utils.py
@@ -2,6 +2,7 @@ import asyncio
 import contextlib
 import functools
 import math
+import os
 import time
 from typing import Optional
 
@@ -287,7 +288,11 @@ async_all_storage = pytest.mark.parametrize(
         ),
         pytest.param(
             "async+redis://localhost:7379",
-            {},
+            {
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                )
+            },
             lf("redis_basic"),
             marks=pytest.mark.redis,
             id="redis",
@@ -308,14 +313,22 @@ async_all_storage = pytest.mark.parametrize(
         ),
         pytest.param(
             "async+redis+cluster://localhost:7001/",
-            {},
+            {
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                )
+            },
             lf("redis_cluster"),
             marks=pytest.mark.redis_cluster,
             id="redis-cluster",
         ),
         pytest.param(
             "async+redis+cluster://:sekret@localhost:8400/",
-            {},
+            {
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                )
+            },
             lf("redis_auth_cluster"),
             marks=pytest.mark.redis_cluster,
             id="redis-cluster-auth",
@@ -328,6 +341,9 @@ async_all_storage = pytest.mark.parametrize(
                 "ssl_keyfile": "./tests/tls/client.key",
                 "ssl_certfile": "./tests/tls/client.crt",
                 "ssl_ca_certs": "./tests/tls/ca.crt",
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                ),
             },
             lf("redis_ssl_cluster"),
             marks=pytest.mark.redis_cluster,
@@ -335,7 +351,12 @@ async_all_storage = pytest.mark.parametrize(
         ),
         pytest.param(
             "async+redis+sentinel://localhost:26379/mymaster",
-            {"use_replicas": False},
+            {
+                "use_replicas": False,
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                ),
+            },
             lf("redis_sentinel"),
             marks=pytest.mark.redis_sentinel,
             id="redis-sentinel",
@@ -365,21 +386,33 @@ async_moving_window_storage = pytest.mark.parametrize(
         ),
         pytest.param(
             "async+redis://localhost:7379",
-            {},
+            {
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                )
+            },
             lf("redis_basic"),
             marks=pytest.mark.redis,
             id="redis",
         ),
         pytest.param(
             "async+redis+cluster://localhost:7001/",
-            {},
+            {
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                )
+            },
             lf("redis_cluster"),
             marks=pytest.mark.redis_cluster,
             id="redis-cluster",
         ),
         pytest.param(
             "async+redis+cluster://:sekret@localhost:8400/",
-            {},
+            {
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                )
+            },
             lf("redis_auth_cluster"),
             marks=pytest.mark.redis_cluster,
             id="redis-cluster-auth",
@@ -392,6 +425,9 @@ async_moving_window_storage = pytest.mark.parametrize(
                 "ssl_keyfile": "./tests/tls/client.key",
                 "ssl_certfile": "./tests/tls/client.crt",
                 "ssl_ca_certs": "./tests/tls/ca.crt",
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                ),
             },
             lf("redis_ssl_cluster"),
             marks=pytest.mark.redis_cluster,
@@ -399,7 +435,12 @@ async_moving_window_storage = pytest.mark.parametrize(
         ),
         pytest.param(
             "async+redis+sentinel://localhost:26379/mymaster",
-            {"use_replicas": False},
+            {
+                "use_replicas": False,
+                "implementation": os.environ.get(
+                    "ASYNCIO_REDIS_IMPLEMENTATION", "redispy"
+                ),
+            },
             lf("redis_sentinel"),
             marks=pytest.mark.redis_sentinel,
             id="redis-sentinel",
